refactor(models,core,ui): migra monetários para Decimal e redesenha navegação
Lote 1 — Modelos: todos os campos monetários (amount, balance, limit_total,
total_amount, current_price, avg_price, etc.) migrados de float para
decimal.Decimal. Validadores Pydantic e propriedades calculadas atualizados.
Invariante: nunca Decimal(float); sempre Decimal(str) ou Decimal("literal").

Lote 2 — Core engines: analytics, behavioral, installment_engine e
weekly_report operam em Decimal end-to-end. Fronteiras com motores de
score (m1_quant, fragility, m2_governance, liquidity) usam float(inv.*)
para manter aritmética de razões/percentuais em float. sum() com acumulador
Decimal(0) em todos os geradores monetários.

UI — Navegação: st.sidebar eliminado em todas as páginas; substituído por
st.columns([1, 4]) com rail CSS. Strip operacional no Dashboard
(Patrimônio / Liquidez / Fragility / Comprometimento / Risco).
CSS nativo do Streamlit oculto; page_link customizado mantido.

Testes: 338 passando em 5 s. pytest.approx removido de asserções monetárias;
substituído por igualdade exata Decimal. test_sidebar_navigation.py adicionado.

+++ models/transaction.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import date as Date
+from decimal import Decimal
 from enum import Enum
 
 from pydantic import BaseModel, Field, field_validator
@@ -44,7 +45,7 @@ class TransactionStatus(str, Enum):
 class Transaction(BaseModel):
     id: str = Field(default_factory=lambda: str(uuid.uuid4()))
     date: Date
-    amount: float
+    amount: Decimal
     type: TransactionType
     category: Category
     notes: str = ""
@@ -56,7 +57,7 @@ class Transaction(BaseModel):
 
     @field_validator("amount")
     @classmethod
-    def amount_must_be_positive(cls, v: float) -> float:
+    def amount_must_be_positive(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError(f"Valor deve ser positivo, recebido: {v}")
         return round(v, 2)
